+++ src/main/java/com/floney/floney/book/service/BookLineService.java
@@ -3,7 +3,6 @@ package com.floney.floney.book.service;
 import com.floney.floney.book.dto.process.DayLines;
 import com.floney.floney.book.dto.request.AllOutcomesRequest;
 import com.floney.floney.book.dto.request.ChangeBookLineRequest;
-import com.floney.floney.book.dto.request.CreateLineRequest;
 import com.floney.floney.book.dto.response.BookLineResponse;
 import com.floney.floney.book.dto.response.MonthLinesResponse;
 import com.floney.floney.book.dto.response.TotalDayLinesResponse;
@@ -12,7 +11,7 @@ import java.util.List;
 
 public interface BookLineService {
 
-    BookLineResponse createBookLine(String currentUser, CreateLineRequest request);
+    BookLineResponse createBookLine(String currentUser, ChangeBookLineRequest request);
 
     MonthLinesResponse showByMonth(String bookKey, String date);
