+++ src/test/java/com/floney/floney/fixture/BookLineFixture.java
@@ -1,7 +1,7 @@
 package com.floney.floney.fixture;
 
 import com.floney.floney.book.dto.process.BookLineExpense;
-import com.floney.floney.book.dto.request.CreateLineRequest;
+import com.floney.floney.book.dto.request.ChangeBookLineRequest;
 import com.floney.floney.book.entity.Book;
 import com.floney.floney.book.entity.BookLine;
 import com.floney.floney.book.entity.BookUser;
@@ -17,8 +17,8 @@ public class BookLineFixture {
 
     public static LocalDate LOCAL_DATE = LocalDate.of(2023, 10, 22);
 
-    public static CreateLineRequest createOutcomeRequest() {
-        return CreateLineRequest.builder()
+    public static ChangeBookLineRequest createOutcomeRequest() {
+        return ChangeBookLineRequest.builder()
                 .bookKey(BOOK_KEY)
                 .flow("지출")
                 .asset("은행")
@@ -27,8 +27,8 @@ public class BookLineFixture {
                 .build();
     }
 
-    public static CreateLineRequest createIncomeRequest() {
-        return CreateLineRequest.builder()
+    public static ChangeBookLineRequest createIncomeRequest() {
+        return ChangeBookLineRequest.builder()
                 .bookKey(BOOK_KEY)
                 .flow("수입")
                 .asset("은행")
